TelegramBotService/requirements.txt
aiogram==3.18.0
# kafka==1.3.5
# kafka-python==2.0.5
python-dotenv==1.0.1
asyncio==3.4.3 
aiokafka==0.12.0
tabulate==0.9.0